Split virtualization tests in two
One for base virtualization tests and one for simple mode tests.
